requirements.txt
gradio
google-generativeai
torch
torchvision
transformers
langchain
langchain-community
langchain-core
PyMuPDF
faiss-cpu
Pillow
python-dotenv
python-docx
